+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/ResponseFormatResolver.java
@@ -2,6 +2,8 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.web;
 
+import java.util.function.BiFunction;
+
 import javax.ws.rs.container.ContainerRequestContext;
 
 /**
@@ -9,15 +11,14 @@ import javax.ws.rs.container.ContainerRequestContext;
  * takes the format name string from URI and a ContainerRequestContext object from which the header Accept field is
  * accessible.
  */
-public interface ResponseFormatResolver {
+public interface ResponseFormatResolver extends BiFunction<String, ContainerRequestContext, String> {
     /**
-     * Resolve desirable format from URI and ContainerRequestContext. In case of a null return, Fili handles null
-     * format at {@link FiliResponseWriterSelector}, which use json as default value.
+     * Resolve desirable format from URI and ContainerRequestContext.
      *
      * @param format  The format String from URI
      * @param containerRequestContext  ContainerRequestContext object that contains request related information
      *
-     * @return A resolved format decided by the function
+     * @return A resolved format decided by the function, null if none match
      */
-    String accept(String format, ContainerRequestContext containerRequestContext);
+    String apply(String format, ContainerRequestContext containerRequestContext);
 }
