+++ fili-core/src/main/java/com/yahoo/bard/webservice/web/endpoints/MetricsServlet.java
@@ -119,7 +119,7 @@ public class MetricsServlet extends EndpointServlet {
 
             MetricsApiRequest apiRequest = new MetricsApiRequest(
                     null,
-                    formatResolver.accept(format, containerRequestContext),
+                    formatResolver.apply(format, containerRequestContext),
                     perPage,
                     page,
                     metricDictionary,
